CFIN-337 Add BASE_URL to constants object, rename file and use BASE_URL in tests
