+++ src/constants/UrlAndParameters.js
@@ -1,0 +1,8 @@
+const URL_AND_PARAMETERS = {
+    BASE_URL: "https://www.york.ac.uk/search/",
+    COLLECTION: "york-uni-courses",
+    FORM: "course-search",
+    PROFILE: "_default_preview",
+};
+
+module.exports = URL_AND_PARAMETERS;
