--- src/constants/UrlParameters.js
@@ -1,7 +1,0 @@
-const URL_PARAMETERS = {
-    COLLECTION: "york-uni-courses",
-    FORM: "course-search",
-    PROFILE: "_default_preview",
-};
-
-module.exports = URL_PARAMETERS;
